+++ mobile/android/android-components/components/service/glean/src/main/java/mozilla/components/service/glean/net/HttpPingUploader.kt
@@ -68,7 +68,7 @@ internal class HttpPingUploader : PingUploader {
         val request = buildRequest(path, data, config)
 
         return try {
-            performUpload(config.httpClient, request)
+            performUpload(config.httpClient.value, request)
         } catch (e: IOException) {
             logger.warn("IOException while uploading ping", e)
             false